Move indexing code from IndexComponent to Index
See https://github.com/hbz/lobid-gnd/issues/213
